Menambah endpoint untuk delete

+++ BookStore-API/DTOs/BookDTO.cs
@@ -14,7 +14,7 @@ namespace BookStore_API.DTOs
         public string Isbn { get; set; }
         public string Summary { get; set; }
         public string Image { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public int? AuthorId { get; set; }
         public virtual AuthorDTO Author { get; set; }
     }
@@ -29,7 +29,7 @@ namespace BookStore_API.DTOs
         [StringLength(500)]
         public string Summary { get; set; }
         public string Image { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         [Required]
         public int? AuthorId { get; set; }
     }
@@ -43,6 +43,6 @@ namespace BookStore_API.DTOs
         [StringLength(500)]
         public string Summary { get; set; }
         public string Image { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
     }
 }
